sapphire_llm/requirements.txt
absl-py==1.4.0
chromadb==0.3.21
google-cloud-bigquery==3.10.0
google-cloud-core==2.3.2
google-cloud-aiplatform==1.25.0
langchain==0.0.146
looker-sdk==23.6.0
PyYAML==6.0
